feat: improve hidden buttons warning
