+++ web/src/components/map/Map/Map.consts.tsx
@@ -38,3 +38,4 @@ export enum BrowseType {
 export const maxZoom = 18;
 export const hexagonSizeZoom = 14;
 export const onMarkerPositionChangeZoomTo = 10;
+export const showMarkersZoom = maxZoom - 2;
